Addressed Memory leak GLFWWindow.getSize and changed the return type Vector2i in order to avoid unnecessary casting

+++ JGL/src/test/java/org/nebula/CameraTest.java
@@ -2,6 +2,7 @@ package org.nebula;
 
 import org.joml.Matrix4f;
 import org.joml.Vector2f;
+import org.joml.Vector2i;
 import org.nebula.jgl.Camera;
 import org.nebula.jgl.data.Shader;
 import org.nebula.jgl.data.buffer.Buffer;
@@ -83,7 +84,7 @@ public class CameraTest {
     }
 
     private void drawTriangle() {
-        final Vector2f winSize = window.getSize();
+        final Vector2i winSize = window.getSize();
 
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
         glViewport(0, 0, (int) winSize.x, (int) winSize.y);
